Copied GraphicsState to C#

+++ library/GraphicsState.h
@@ -39,16 +39,15 @@ typedef struct
   unsigned short VerticalOffsetRegister;
   unsigned short VPitch;
 
-  unsigned int DistoOffset;
-  unsigned int NewStartofVidram;
-  unsigned int StartofVidram;
-  unsigned int VidMask;
+  unsigned long DistoOffset;
+  unsigned long NewStartofVidram;
+  unsigned long StartofVidram;
+  unsigned long VidMask;
 
   unsigned char  BorderColor8;
   unsigned short BorderColor16;
-  unsigned int   BorderColor32;
+  unsigned long  BorderColor32;
 
   unsigned char  Lpf[4];
   unsigned char  VcenterTable[4];
-
 } GraphicsState;
